refactor: Restructure event handling by consolidating handlers and removing unused files
